Improve payment duplicates check
This patch corrects some glaring errors in the payments duplicates check
that has been causing the exits specifically to accept all payments
twice and give double credit.

The crux of the issue is that when the function to_validate was called
it did not check payments currently waiting to be verified.

This is probably because the to_verify list is a hashset, so it would in
the past automatically reject duplicates. But at some point an Instant
for the time we recieved the payment was added. This did not produce any
immediate problems, but many months or even years later make_payments_v2
was introduced, this new payments endpoint had routers play back their
full local payment history to ensure the other router was in sync
despite reboots/crashes.

This means that a new payment, with a differnet value for the instant
recieved could be inserted into the to_validate list if for some reason
a second call of make_paymetns_v2 was called within a short amount of
time.

This case happens to be very common on the exits during high bandwidth
periods. One tx will go through and within less than a minute another
one will go through as well. But thanks to make_payments_v2 the last
payment was inserted with a new instant when the next payment was also
being added. Resulting in a duplicate and double credit.

In order to remove this issue several steps have been taken

1) an explicit check of payments in flight has been added to to_validate
2) The ToValidate struct now has a custom implementation of hash which
   only inspects the txid, meaning the hashset duplicate prevention
   works again. This is a redundant check in case (1) fails
3) A special error type for duplicate payment has been added which is
   then used by two new unit tests to confirm that (1) remains
   functional