Use combined txid for Althea payments
This patch uses a combined txid for Althea payments and xdai paymetns
this removes an extra field to check and optionality from the struct.
Also in this change the txid field in paymentx is now mandatory with a
new type UnpublishedPaymentTx taking it's place in situations where it
would be none.

This helps enforce correctness via types and generally prevent
accidents.