Only have one tunnel per peer ip
This is the first major change to the tunnel opening and management
logic in 4 years. Several key things have changed during that time.

First and foremost native futures have reduced the scoping dramatically
letting us simplify the tunnel manager logic by adding member functions
and relying more heavily on the borrow chcker.

Second, this change is modifying the actual operation of tunnel manager
by changing the criteria for opening a new tunnel to be a unique wg_key
and ip address, whereas before it was a unique wg_key, ip address, and
interface index.

The interface index is a concept directly tied to ipv6 link local
addresses which are only valid for a given interface. We would always
end up with two tunnels one for ifidx 0 (listening) and one for a
specific ifidx (for reaching out to the peer). Once wireguard makes
contact it abstracts this away for us anyways, but in the past we
figured better safe than sorry as far as tunnel formation.

What's lead to changes now is that the number of extra interfaces is a
significant performance problem, babel does not handle many neighbors
well and having two connections per peer doubles the number babel has to
deal with.

The case where having a ifidx zero connection is useful is when the peer
router has for example a cable moved one port to the left or right. If
our router has only a tunnel with a specific ifidx the tunnel would
never recover. But with an ifidx zero connection it would resume.

As it happens in the past 4 years we've encountered and solved this same
issue in antoher way, tunnel gc will delete and re-crate a tunnel in
this case, allowing for recovery in this port switching case, although
on a not ideal timespam of 15 minutes rather than a few seconds.

Given our babel performance problems, running tunnel manager gc faster
will actually be easier than dealing with babel performance. Hence this
change.