Enable /tmp/ logging on machines with sufficient memory
This patch updates compressed_log to enable logging to files in /tmp
only on machines with memory to spare, this effectively means only x86
routers and some models with 512mb or 1gb of memory.

The hope is that with this change we can track down obscure bugs without
pushing the burden of debugging overhead to the most constrained
devices.