This commit is the first step in removing the postgreSql database from exits
The goal of this change is to make it possible for many thousands of exits
to easily run off a single database with no shared credentials and also to
make it possible for many different organizations to cooperate on this database.

In order to achieve this we will be replacing the postgresql database currently
used by the exits with a smart contract which will be used to register users

Client register flow has changed from client -> Exit -> Ops. Exit forwards a request to an external server with sms api key in order to
verify a client. That server will have write access to the smart contract to add the client, and then return their registration status to
exit, which then forwards that request back to the client