Add base solidity directory for Althea-L1 contracts
The solidity folder functions as a solidity development environment
under Rita, with functions to write + unit test + compile solidity
contracts for deployment on the Althea-L1 EVM.

This folder is a simpler version of the one in
github.com/althea-net/althea-L1. For contracts to be available on the
Althea-L1 EVM in integration tests, the testing startup process can
execute solidity/scripts/contract-deployer.sh

See the README.md for lots of info and the commented files for
explanations of how to deploy and test new contracts.