Sort when converting between flat and map usage data
This commit contains several quality of life fixes and one correctness
fix to the recent commit updating usage tracker. The correctness fix is
that we did not sort data when converting from the new back to the old usage
tracker format. Since we kept the old format for all use cases involving
frontends and graphing this would have caused all the graphs / data to
be incorrect when displayed.