Add region codes to exits in exit db contract
This commit adds region codes to the exit db contract. The intent of
this feature is to allow routers to avoid connecting with exits that are
not compatible with their region. Otherwise we risk routers roaming to
exits that can not support them. This rejection process could have been
done exit side instead of client side but that would add a period of
downtime for the user while the client did region negotiation with the
exit. Not to mention the load on the exit of doing many region
negotiations at once.