Add retry system for withdraw checks
In order to reduce load on the xdai nodes we recently set more
restrictive limits on queries. Apparently the 10k block limit here
exceeds that, this change both reduces the limit and adds a retry system
to help make the function call more reliable when needed.

Finally we reduce the size of the search window to a mere hour rather
than a week. Since the number of required log requests was inordinate.