This commit updates the client exit manager to work with the new exit format
i.e. A vector of individual exits as potential exits to an exit can connect to, rather than
subneting or clustering of exits that was done in the past. This allows exits to have only one
identity that the client cares about, which updates gateway billing to use only one exit id in
debt keeper

The Exit list now returns a list of exit id with their respective allowed regions to connect to and
payments accepts, the client (which has these fields set up in its config) will now filter through all
exits it receives, put the ones it could potetially connect to in the ExitServer list. This allows us to
query for general details of these exits. If we are already registered to one exit, we are registered to all
exits, so these new exits move from New -> Registered, which then allows us to switch to them easily when
an exit goes down.