Auto generate config on start if file is not found
This patch removes cruft around startup args for Rita and modifies the
default configuration file behavior to generate the config if one is not
found.

Previously I didn't think generating a config was valuable as a feature.
But recently I've see strange behavior on some devices with limited
storage where the config was no longer present.