Fix exit ctrl-c handler
This patch includes a few cleanups for the exit, first off we fix the
ctrl-c handler which means the exit restarts gracefully when you ask it
to do so. Essential behavior.

Next we remove hte --future agrument that has never been used and
finally we change the internal config passing to Pathbuf to match rita
client