Fail fast if exit database contract can't be queried
This commit modifies the way the exit handles users list queries to be
in line with the operating philosophy we usually use. Which is to crash
quickly if the initial configuration or state is not correct. But one
running the system should remain running if at all possible.

Previously we tried to query the users list for 5 minutes before
crashing. This is kinda the worst of both worlds, it won't be caught by
a deployment script just checking for quick crashes and it also takes
down the application any time we lose connection to the full node for
more than 5 minutes

The new design ensures that an already running exit will ride through,
and that the program wil crash quickly if not configured properly on
startup.