Lock free payment controller refactor
As it turns out eliminating the global locking structure in payment
controller was pretty trivial. The only data being passed into the
module via the lock was actually being generated in the same thread in
the first place.

In addition to removing the lock the retry mechanism has been adjusted
to use the same submission flow as the normal payment flow. The old rety
flow depended on make_paments v1 only and was generally just code
duplication.