Remove nonce, net_version, and gas_price from oracle
When the blockchain oracle was first developed we where using raw
transaction handlers to send tx, now we're using the web30
send_transaction endpoint which can easily grab all this data on the
fly.

Likewise we used to need to keep track of nonces becuase ETH tx
validation could be very slow, so slow that we wanted to increment the
nonce and queue the next transaction. This is not the case post EIP1559
and either way we are using Gnosis chain or Althea L1 for payments that
generally also do not have that problem. So this functionality is mostly
vestigial.