Flatten open_tunnel logic
This particular corner of the open tunnel logic affects how exactly we
open tunnels for LTE clients and gateways in odd ways. Most tunnels are
peer to peer tunnels and follow the ipv6 link local address flow. But
some are exit tunnels which require specific handling of the default
route and some are LTE tunnels or just ipv4 peers over local networks
(same thing).

A recent change in the commit 'make tunnel opening more robust' tried to
exit this function early in the case that we couldn't lookup a gateway
interface. But the gateway interface was only required due to the
socket_addr_to_string function which itself only consumed the interface
in the ipv6 case.

Once the logic is flattened and the socket to string function removed it
becomes clear that we are getting irrelevant info just to feed it to a
function that throws it away and our error cases are not consistent.

The new logic is both shorter and more correct.