Do not requeue failed transactions in payment_controller
When refactoring payment controller and debt keeper an oversight was
made. Payments have never been requeued to try again in the payment
controller module. But I added a piece of logic that did this,
duplicating payments that failed due to any issue.

This patch removes that incorrect logic and adds a check in debt keeper
that ensure that payments can not succeed or fail more than once by
asserting the in flight status.