Set babeld settings on each interface
When babeld settings where overhauled we attempted to set the default
for new interfaces and simply rely on that to be carried forward across
all interfaces. This does not seem to work the way we could expect at
least for the rtt timestamps feature which is essential to latency
computation.

In my tests setting these values per interface works reliably so that's
what we'll do going forwward