Move self rescue tools to their own thread, add tunnel activity check
This patch moves the self rescue ping check to it's own thread, where it
won't be subject to problems in the rita client thread. A new check has
also been added that checks for inactive tunnels, this handles observed
cases where routers would be online and pinging find, but not opening
new tunnels for clients.