Simplify usaing saving loop
there where several unused values in this loop and the config option for
the save interval complicated matters further. This change means that
saving intervals are very predictable.

Due to the default config value all routers where waiting 48 hours
between saves. Which does have negative consequences when a payment is
made but not saved in debt keeper.