Fix: unified timeout logic
The outgoing payment succeeded message was not migrated to it's new location
properly in the xdai case. And the incoming and outgoing where not
migrated in the Althea l1 case. This would have caused a double credit
for payment on Althea L1 and a panic for double sending a payment.

But on xdai since the payment succeeded logic is still called only once
everything should operate correctly. Not that we're going to let this
float around long to find out.