Update self rescue code to avoid reboot loops
This patch helps avoid reboot loops in a very specific edge case where a
router is always updating it's system clock on power on.

The fact that this case exists reveals one of the several flaws with
this strategy of integrity check, and it's probably better to evolve the
pattern of the various rita loops with thier watchdog threads such that
the watchdog thread can watch a local status report from the inner
thread and restart as needed.

This would eliminate the need for global state to be involved in the process
of determining if these loops are healthy.