Rita exit waits on startup
Previously Rita exit would halt immediately if a number of checks did
not pass. The system checks went beyond the basic config sanity check
and verified that the exit was ready to run and grab the active users
list from the exit contract.

Halting if any of this is configured incorrectly is great behavior for
the Ansible script deployed exits as it immediately indicates the issue
to the operator.

But for the firmware based exits simply halting the process will be
impossible to debug or identify. So we provide a new flow that will give
similar vibility to the firmware based exit operators by having the
checks continue to run in a loop and post data to rpc endpoint from
which it can be displayed on the dashboard.