Move WiFi settings over to Rita common
This will allow the exit firmware images to have working wifi to go with
their normal lan functions. It seems a little unintutive that people
would want wifi on a dedicated exit appliance but it's usefulf or both
servicing the device if required and generally becuase people often want
to provide free wifi at a gateway.