Fix: WiFi radios without the wlanX tag
Openwrt internally points at radios as radio0,1,2 and so on. But
traditonially these interfaces where labeled wlan0,1,2 and this was the
default several versions ago. This behavior and naming is no longer the
default but can be replicated with naming tags in the config which are
optional.

The newer linksys routers where shipped with a config that does not
contain these optional tags, so this patch modifies our parsing logic to
accept untagged configs. This is backwards compatible.