Ensure tunnel deletion actually finishes before startup
The previous solution did not account for ensuring the tunnels where
actually deleted and instead just launched a lot of background
operations that where immediately disrupted by the creation of new
tunnels since we didnt' wait for deletion to finish.

This is only apparent on machines with such a large number of tunnels
that deletion couldn't finish before the program started in earnest.

When Rita is crashing this leads to a feedback loop where rita crashing
and restarting leaves more and more cruft tunnels behind making the
machine less responsive generating the next crash.

Also included here are a some logging refinements.