Fix: Ensure tunnel deletion actually happens before startup
For truly uncomprehensible reasons on some systems, a dynalink wrt-36x
in this case changing from parallel tunnel deletion with no join to
parallel tunnel deletion with a join results in some interfaces being
left behind even when they where not before.

I believe this may have something to do with the timing of the
instructions executing. Regardless failure to delete old tunnels results
in strange behavior and signifncant packet loss as the expired tunnel
sticks around and sucks in valid packets that are then not routed on
since babel is not attached.

We can't simply return to prallelized deletion without join becuase on
some other hosts (x86 gateways with huge numbers of tunnels) it doesn't
complete in a reasonable timeframe which also results in stray tunnels,
packet loss and other bad behavior.

We can't do a fully sync tunnel deletion either becuase on some systems
deleting all the tunnels syncronously will take tens of minutes an
unacceptable amount of downtime.

So we use this compromise solution where we first delete in parallel
then deploy a syncronous delete to handle the stragglers. This has been
tested to work on both the small and large scale and provides a fast
restart time even with hundreds of tunnels without compromising on
accuracy.