Replace old checkin flow with websocket loop
This new and improved version of the checkin loop no longer needs to
wait for ops to send responses to http requests in order to complete
the checkin flow and receive operator updates. Now, with the websocket
replacement loop we can be sure ops has received data and not block on
whether or not we receive an OperatorUpdateMessage as ops will process
separately and send back on its own time.

The websocket loop will keep trying to connect to ops even if connection
is dropped or refused (as in cases where ops restarts)