Begin refactoring exit client code
This commit is the first checkpoint in a full refactor of exit client
code.

Starting by changing the settings structure to be more appropriate for
the global curated list of exits maintained by the exit registration
smart contract rather than many distinct exit clusters populating the
config.

The next phase of this change will be to dig into the code responsible
for for getting the exits list and modify it to work with a larger in
memory list / update the config with new exits.