Remove exit blacklist feature
The exit blacklist feature was originally designed for the case where we
used subnets to identify exits. Meaning any device in a given configured
subnet would be assumed to be an exit. This strategy was dropped for
it's many deficiencies and with the new solidity contract based
structure we're moving to it's no longer a risk. If we can't
successfully checkin an decrypt an exit response we're not going to
select the exit anyways and there's absolutely no risk of a 'rouge exit'
sneaking in.