Remove bracketing error in exit loop
This one line patch resolves an issue introduced 10 months ago in the
patch "This commit updates the client exit manager to work with the new
exit format". That split off the exit status updates and slacc setup
from the main exit loop by accident. The only reason this has not been
noticed is because this code never entered production since that time.