Cleanup exit loop
This patch cleans up the exit loop by separating it out into several
functions.

The new code organization revealed several itneresting bugs, including
two different indentation errors. One corrected in the previous commit
and another that would cause the loop to spinlock on settings if we had
no selected exit yet.

In addition to that some flow problems have been modified. Since we
always have a selected exit when we go into the exit switcher we now
attempt to setup that exit right away rather than waiting for a
successful run of the exit switching logic. Worst case scenario we do
end up having to wait but best case we can have a tunnel up much faster.

Note that because we're always saving out to the config the best case is
the common case and on average this can speed up reboot times for users
by 30 seconds or more.