Convert to a single registration state var
This patch converts the exit client state from expecting an array of
distinct exit clusters to expecting only a single exit cluster active at
any given time as defined by the current exit registration smart
contract.

Bootstrapping exits in the devices local config are used to get an
update of the smart contract state for roaming exits. This is one part
of a still in progress refactor.