Websocket send/receive structs cleanup
We now send websocket messages as a variant of an enum type to make
deserialization easier. This change also groups related(processed
together on ops) data together to simplify and reduce ops db load,
as it doesn't make sense to send individual messages for every
piece of data when a DB read/write to a device entry is needed for
processing.

Previously used, now-legacy checkin and update structs have been
kept but moved to the legacy.rs file to allow ops to retain function
for older routers.