Overhaul exit switcher, region management, and exit selection
This patch represents a continued overhaul of the exit manager
functionality. Two major things are occuring a simplification and an
overall change in the functionality of the code to track a single
cluster of exits managed by a smart contract instead of many clusters of
exits managed by ip ranges.

This implies a lot of touch point changes, to exactly how logic changes.
This makes up a significant part of this commit.

The biggest single change in this commit is a total re-write of the exit
switching functionality which was designed for subnet based exits
originally and inherited complexity that is no longer needed given our
new reality of exits with a fixed list of members. The new simplified
design maintains the vast majority of the old feature set but does not
for example implement the blacklisting feature, where exits that where
online but enough to be in babel but otherwise not working correctly are
removed from the selection process.