Temporary fix for multi exit test
This is a temporary fix for the multi exit test that has exit clients
use the config entierly for the next exit to select, the next step in
our refactor is to refactor the exit list component that's failing here.
So I wanted to get the test green without fully updating that component.