Break up interop.rs
This is a long overdue change to completely break up and remove the
interop.rs file in althea types. Which has from it's creation been a
dumping ground for structures that we need to use across multiple
crates. Obviously throwing any struct you use across crates into one
file for years on end is a great way to create a very long and
unreadable file.

This patch finally splits it up into logically grouped files that should
be much easier to follow.