Tag specific tests to run serially
This commit adds a new test dependency that allows us to tag which tests
need to be run serially versus in parallel. This lets us parallelize the
rest of our test suite and run serially only those tests that must be run
serially.

This dramatically speeds up the tests because we have a number of
fuzzing and other longer run tests that take a few minutes.