Make SignedExitList more robust to signature breaking
This patch centers around making the SignedExitList generally more
robust. By making the individual vars private with only getters, and
adding a signer field to the struct itself it's generally impossible to
accidentially modify the data and mess up a signature. Or to construct a
signed exit list when it's not needed.