Unify actix deps, remove -async name prefix, do not use rustls
This patch removes the -async naming prefix for actix dependencies that
we adopted when pre-async and async dependencies had to co-exist in the
same binary.

While doing this all actix deps and openssl are unified into workspace
dependencies. Making them easier to track and manage.

Finally we use openssl instead of rustls for ssl in the exit trust root
as it turns out that any use of rustls poisons the workspace for easy
cross complilation.