Fold registration server functions into the exit trust root
When we first moved to the exit db smart contract the registration
server was created to provide a way to register bootstrapping clients
despite them having no balance and even no internet connection, by
passing that data to exits.

As this design has gotten more general and the requirements have
expanded to support out of the box appliance exits the root of trust
server was created, it's natural to fold the registration server info
the reference root of trust implementation in this repo.

As part of this process the trust root, and the api for the client
registration has been refactored, simplified, and cleaned up.