Add to_bytes to identity
The get_hash and get_hash_array functions are unstable between rust
versions due to changes in the internal hashing function. This patch
adds deprecation tags to these functions and implements a to and from
bytes representation for the Identity struct.

This is distinct from the Ethereum ABI functions, as the byte
representation is entierly defined by the Identity struct and not the
Ethereum standard. As such it's a lot more straightforward to implement.

The downside is that the new implementation is quite a bit larger than
the old hash implementation at 68 bytes versus 8 bytes. But that's not
really a big deal, for 5000 routers thats an extra 300MB of data used in
indexing. Furthermore there's no risk of someone trying to generate an
Identity with a hash collision.