Simplify internal and ipv6 assignment
This patch modifies the internal ip assignment and ipv6 address
assignment to match in design and test coverage with the external ip
assignment code.

This simplification is mostly achived by removing the requirement that
clients get the same ip each time they connect even between reboots.
Reviewing the code I decided it was easier to fix that problem on the
rita_client side than to deal with the solution here.