Fix: Unify client and exit wg port constants and literals
Since the exit registration smart contract does not store data about the
client wireguard listen port it exists only as a constant. Before this
patch it was two constants one in the exit code and one in the client.

It's not strictly required to set the client listen port correctly, the
server listen port is passed through the registration contract and is
sufficient to generate a handshake. But obviously there's no reason to
have two constants for the same value around.

The issue that actually generated errors was in the integration test,
where the wg_exit listen port for smart contract registration was
hardcoded indepenently of the default value in the rita_exit settings.
By unifying these constants a confusing integration test error is no
longer possible.