SNAT routing: corrected setup and teardown logic + cleanup
The previous implementation of setup would have readded the same SNAT
rules to nftables any time a new client registered to the exit, which
would clog the tables and mess with teardown logic. In this fix we
only add rules when a client is not in the external ip assignments
list AND also online as defined by WG_INACTIVE_THRESHOLD. When a client
is deemed offline they are now removed from external ip assignments
list as they should have been on top of just having their snat rules
torn down.