Backport unified timeout logic to beta20 branch
This is a backport of the commits "Unified payment timeout logic, add
config file validation" and "Fix: unified timeout logic" to the beta20
branch namely removing things that don't exist in this branch like the
Althea L1 transaction platform support.

This change overhauls how timeouts are handled unifying them in
payment_validator rather than in debt keeper, removing the lock based
message passing involved and adding in lots of little changes as far as
validation.

This is not a full backport of all the nice associated features, but a
scoped backport of just what's needed to get the new flow working on
this branch.